TL commit existing

+++ Gitpro2/src/main/java/com/nt/Gitpro2/Airthematic.java
@@ -3,11 +3,11 @@ package com.nt.Gitpro2;
 public class Airthematic {
 
 	
-	public int sum(int x, int y) {
+	public double sum(double x, double y) {
 		return x+y;
 	}
 	
-	public int sub(int x,int y) {
+	public double  sub(double x,double y) {
 		return x-y;
 	}
 }
